refactor: organize the units into different files
issue #14
